rename src/SharedComponents/* to src/components/*
