+++ src/Screens/HomeScreen/HomeScreen.jsx
@@ -1,12 +1,12 @@
 import React from 'react'
 
-import Button from "../../SharedComponents/Button/Button"
-import Top from "../../SharedComponents/Top/Top"
-import Title from "../../SharedComponents/Title/Title"
-import Card from "../../SharedComponents/Card/Card"
-import Screen from "../../SharedComponents/Screen/Screen"
-import Bottom from "../../SharedComponents/Bottom/Bottom"
-import Footer from "../../SharedComponents/Footer/Footer"
+import Button from "../../components/Button/Button"
+import Top from "../../components/Top/Top"
+import Title from "../../components/Title/Title"
+import Card from "../../components/Card/Card"
+import Screen from "../../components/Screen/Screen"
+import Bottom from "../../components/Bottom/Bottom"
+import Footer from "../../components/Footer/Footer"
 import "./HomeScreen.css"
 
 function Or() {
@@ -29,7 +29,7 @@ function HomeScreen() {
         <Title text="A party game for horrible people"/>
         <Button text="Join Party" className="center" link="/join-party"/>
         <Or />
-        <Button text="Create Game" className="center" link="gameplay"/>
+        <Button text="Create Game" className="center" link="/gameplay"/>
         <Footer>
           Like us on <a href="http://www.facebook.com/yusufameri"> Facebook!</a>
         </Footer>
