+++ src/Screens/StartGameScreen/StartGameScreen.jsx
@@ -1,49 +1,41 @@
 import React from 'react'
 
-import Screen from "../../SharedComponents/Screen/Screen"
-import Top from "../../SharedComponents/Top/Top"
-import Title from "../../SharedComponents/Title/Title"
-import Bottom from "../../SharedComponents/Bottom/Bottom"
-import Card from "../../SharedComponents/Card/Card"
-import Button from "../../SharedComponents/Button/Button"
-import Footer from "../../SharedComponents/Footer/Footer"
+import Screen from "../../components/Screen/Screen"
+import Top from "../../components/Top/Top"
+import Title from "../../components/Title/Title"
+import Bottom from "../../components/Bottom/Bottom"
+import Card from "../../components/Card/Card"
+import Button from "../../components/Button/Button"
+import Footer from "../../components/Footer/Footer"
+import PlayerList from "../../components/PlayerList/PlayerList"
 import "./StartGameScreen.css"
 
-let players = [
-  "Yusuf",
-  "Salman",
-  "Joseph",
-  "Sal",
-  "Albert"
-]
 
-function PlayerList(props) {
-  return (
-    <div className={`${props.className}`}>
-      <ol>
-        { props.players.map((player => <li>{player}</li>)) }
-        <li><input type="text" className="enterName" placeholder="Enter Name Here"/></li>
-      </ol>
-    </div>
-  );
-}
-
-function StartGameScreen(props) {
+class StartGameScreen extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = {
+      players: ["Yusuf"]
+    }
+  }
+  
+  render() {
     return (
       <Screen>
         <Top>
-          <Card cardType="Link" link={props.match.params.partyCode} />
+          <Card cardType="Link" link={this.props.match.params.partyCode} />
         </Top>
         <Bottom>
           <Title text="Players Joined"/>
-          <PlayerList players={players} className="center"/>
+          <PlayerList players={this.state.players} className="center"/>
           <Button text = "Join Party" className="center" />
           <Footer> 
-            {players.length < 3 ? "Need at least 3 players to Join" : "Ready to start Game!"}
+            {this.state.players.length < 3 ? "Need at least 3 players to Join" : "Ready to start Game!"}
           </Footer>
         </Bottom>
       </Screen>
     );
+  }
 }
 
 export default StartGameScreen;
