+++ src/components/DropCardSpace/DropCardSpace.css
